Adding in website files and build scripts
Putting the website in the development branch with a build script so we can update themes, test them with the debug page, see how the site would render, then just build a tarball and update the gh-pages branch with the changes.  This also removes duplicate documentation for the options and helps to keep things up to date.