Bugfix to address issue #251
The div/span should only be hidden if the element itself is hidden.
Previously the div/span were hidden if the element was hidden by itself or
by any parent.  For the case of jQueryUI and multiple tabs, the non-showing
tabs are all hidden and this was causing Uniform to hide the div/span.
When the tab was shown, the div/span would not be unhidden.

I believe this change will fix the problem.